add merging PR section and code of conduct
This adds a section to CONTRIBUTING.md that talks a little bit about merging/reviewing PRs for those who have push access. This also adds a code of conduct. This group was borne of the WeCode event which is intentional about inclusivity and creating a safe environment so it seemed to make sense that this project should do the same.